day17/avro-log-system/requirements.txt
# Core Avro dependencies
avro-python3==1.10.2
fastavro==1.9.4

# Web framework for dashboard
flask==3.0.3
flask-cors==4.0.0

# Testing framework
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# Data handling
dataclasses-json==0.6.6
pydantic==2.7.1

# Utilities
click==8.1.7
colorama==0.4.6
tabulate==0.9.0

# Development tools
black==24.4.2
flake8==7.0.0
